sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atlast-ecp"
version = "0.7.0"
description = "ATLAST Evidence Chain Protocol — trust infrastructure for AI Agents"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "ATLAST Protocol Working Group" }
]
keywords = ["ai", "agent", "trust", "ecp", "atlast", "llm", "audit", "evidence"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
requires-python = ">=3.10"
dependencies = []  # Zero required dependencies — Fail-Open by design

[project.optional-dependencies]
crypto = ["cryptography>=42.0"]   # For ed25519 signing (recommended)
gemini = ["google-generativeai>=0.5"]  # For Google Gemini wrap() support
litellm = ["litellm>=1.30"]  # For LiteLLM wrap() support
otel = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-instrumentation-openai>=0.20",
    "opentelemetry-instrumentation-anthropic>=0.20",
]
proxy = ["aiohttp>=3.9"]          # For atlast proxy / atlast run
langchain = ["langchain-core>=0.1"]   # For LangChain adapter
crewai = ["crewai>=0.28"]            # For CrewAI adapter
all = [
    "atlast-ecp[crypto,proxy,otel]",
]
dev = [
    "pytest>=8.0",
    "anthropic>=0.25",
    "openai>=1.30",
    "cryptography>=42.0",
]

[project.scripts]
atlast = "atlast_ecp.cli:main"
atlast-ecp-mcp = "atlast_ecp.mcp_server:run_stdio_server"

[project.urls]
Homepage = "https://github.com/willau95/atlast-ecp"
Repository = "https://github.com/willau95/atlast-ecp"
Documentation = "https://github.com/willau95/atlast-ecp/blob/main/ECP-SPEC.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["atlast_ecp*"]
